printing the size of variable types

+++ 0x00-hello_world/6-size.c
@@ -10,11 +10,11 @@ int main(void)
 	long int l;
 	long long int L;
 	float f;
-printf("Size of char: %lu byte(s)\n", sizeof(c));
-	printf("Size of int: %lu byte(s)\n", sizeof(i));
-	printf("Size of long int: %lu byte(s)\n", sizeof(l));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(L));
-	printf("Size of float: %lu byte(s)\n", sizeof(f));
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(l));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(L));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
